cloud/pyproject.toml
[project]
name = "cloud"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "redis[hiredis]>=5.0.0",
  "sqlalchemy>=2.0.0",
  "asyncpg>=0.29.0",
  "httpx>=0.27.0",
  "python-jose[cryptography]>=3.3.0",
  "google-genai>=0.8",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "pytest",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
indent-width = 2
target-version = "py311"
lint.extend-select = ["E", "F", "B", "UP", "I"]
fix = true
show-fixes = true

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"
